requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
selenium==4.16.0
requests==2.31.0 